consolidate client to single file and use env vars on windows
