+++ DragonFruit.Six.API.Tests/TestData.cs
@@ -1,12 +1,12 @@
 ﻿// Dragon6 API Copyright 2020 DragonFruit Network <inbox@dragonfruit.network>
 // Licensed under Apache-2. Please refer to the LICENSE file for more info
 
-using System;
 using System.Collections.Generic;
 using DragonFruit.Six.API.Clients;
 using DragonFruit.Six.API.Data;
 using DragonFruit.Six.API.Data.Containers;
 using DragonFruit.Six.API.Enums;
+using DragonFruit.Six.Developer.Clients;
 
 namespace DragonFruit.Six.API.Tests
 {
@@ -36,6 +36,6 @@ namespace DragonFruit.Six.API.Tests
             }
         };
 
-        internal static readonly Dragon6Client Client = new Dragon6TestClient(Environment.GetEnvironmentVariable("devKey"));
+        internal static readonly Dragon6Client Client = new Dragon6DemoClient();
     }
 }
